refactor: Rename Appbar to Navigation to be more encompassing
